Update NoiseGenerationSystem and HeightNoise to use integers over floats to mark exactly where the top layer of the world should be. Update ChunkGenerationSystem to generate negative Y Chunks to always have a nice big underground section

+++ Assets/_Scripts/MapGeneration/World/DOTS/ChunkSystems/NoiseGenerationSystem.cs
@@ -47,7 +47,7 @@ public partial class NoiseGenerationSystem : SystemBase
 
                 float baseNoise = heightNoise.GetNoise(x, z);
                 // normalize the noise value to a range suitable for height
-                float heightValue = (baseNoise + 1f) * 0.5f * WorldConstants.WORLD_HEIGHT;
+                int heightValue = (int)((baseNoise + 1f) * 0.5f * WorldConstants.WORLD_HEIGHT);
                 
                 buffer[index++] = new HeightNoise
                 {
